+++ inst/doc/create_article.R
@@ -11,3 +11,9 @@ library(rjtools)
 ## ----eval = FALSE-------------------------------------------------------------
 #  create_article(name = "quokka-bilby")
 
+## ----eval = FALSE-------------------------------------------------------------
+#  rmarkdown::render("quokka-bilby.Rmd")
+
+## ----eval = FALSE-------------------------------------------------------------
+#  install.packages(c("plotly", "ggplot2", "kableExtra", "palmerpenguins"))
+