+++ inst/sec-depth.lua
@@ -1,7 +1,19 @@
- 
-function Header (h)
-  if h.level > 1 then 
-    h.classes:insert 'unnumbered'
-  end
-  return h
+function Pandoc (doc)
+  -- Create and number sections. Setting the first parameter to
+  -- `true` ensures that headings are numbered.
+  doc.blocks = pandoc.utils.make_sections(true, nil, doc.blocks)
+
+  -- Shift the heading levels by 1
+  doc.blocks = doc.blocks:walk {
+    Header = function (h)
+      if h.level > 2 then
+        h.classes:insert 'unnumbered'
+      end
+      h.level = h.level + 1
+      return h
+    end
+  }
+
+  -- Return the modified document
+  return doc
 end